fix(backend): file upload not returning unsuccessful status codes
Discovered an API route handler bug for `/api/sku/upload`, whereby whenever the a DB SQL run fails, the request handler response does not reflect this, and only returns 200.

Root cause: the db operations are asynchronous, and so wrapped a Promise around the DB operations and awaited the overall operations until the upload fails, or successfully completes.