chore(frontend): refactor dialog trigger component
The component is leaner and more maintainable, by enforcing just a component to be passed in as the dialog trigger.

This hands full control of the underlying trigger component rendering back to the consuming parent component.

Also renamed prop `triggerItem` to `triggerComponent` to be improve clarity of what needs to be passed in.

Finally, added further typescript definition documentation (multi-line comments) for the props of `<DialogTriggerItem>` component.