fix(frontend): toast element region height overlaps full screen height
Whenever a toast notification is rendered, it could cause the toast to override duration and stay on the screen, which only happens when the user hovers or clicks on a region of the UI underneath the toast notification element.

This is because previously, the actual height of the toast element container is preset to the full height of the viewport `max-h-screen` (via TailwindCSS). The solution is to ensure the element container height is maxed up to the size of the element content `max-h-fit`.