fix(frontend): data table column headers spelling errors
Added new unit tests for `<DataTable />`, which also picked up the column headers as being incorrectly spelt 🙈

Data table column header spellings are now corrected, along with unit tests to ensure it is rendered correctly, should any further changes are to be made to the component in future.

Finally, just refactored an exiting unit test for `<SkuAddProductBtn>`, by addding in BDD descriptive comments, with the intention to improve developer experience and maintenance when reading the flow of the unit test. It would also be consistent with other existing unit tests too.